Add end turn logic to features
All steps (except getting sent to jail and breaking out etc) should not
end turn automatically. So require the end turn button to be clicked
before swapping players.

Fixes up all the features to do this, adds a couple more to catch some
edge cases, and adds more steps. And of course adds in the code to do
it.

Uses a notion of "expecting" rolls, and defaults to expecting 2 on turn
change, and adds 2 on a double roll, and subtracts 1 on a single roll.
Will need to expand when utility rent comes in